pyproject.toml
[tool.poetry]
name = "account-managment"
version = "0.1.0"
description = ""
authors = ["Flavio Oria <flavio.oriap@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10"
fastapi = "0.111.0"
bcrypt = "4.1.3"
pyjwt = {extras = ["crypto"], version = "2.8.0"}
python-dotenv = "1.0.1"
emails = "0.6"
jinja2 = "3.1.4"
mjml = "0.11.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.6"}
pydantic = "^2.9.1"
pydantic-settings = "^2.5.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
pre-commit = "^3.8.0"
isort = "^5.13.2"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
